requirements.txt
beautifulsoup4==4.12.3
python-telegram-bot==21.7
redis==5.0.8
Requests==2.32.3
